Test Load-Fix-Validate chain in AasxToolkit
This patch introduces a test case that tests the command chain against
sample XML and JSON data (analogous to ValidateAAS.ps1 debug script).

Additionally, a couple of streams and files were not properly closed
(due to omitted `using` or `finally` blocks), so this patch fixes them
as well so that the tests pass.